refactor(ui): share one byte-size formatter across the panels
The stage HUD, batch converter, debug overlay, and streaming cache readout
each carried their own formatBytes with a different base (1000 vs 1024),
precision, and small-value handling, so the same size could read '12 MB' in
one panel and '11.8 MB' in another. Extract the already-tested formatByteSize
into its own leaf module (so the UI can share it without pulling in loadPlan),
re-export it from preloadSummary for back-compat, and point all four panels at
it. Output is now consistent: binary units, one decimal from KB up.
